+++ src/main/java/com/yohanii/lostandfound/LostandfoundApplication.java
@@ -2,6 +2,8 @@ package com.yohanii.lostandfound;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.web.filter.HiddenHttpMethodFilter;
 
 @SpringBootApplication
 public class LostandfoundApplication {
@@ -10,4 +12,8 @@ public class LostandfoundApplication {
 		SpringApplication.run(LostandfoundApplication.class, args);
 	}
 
+	@Bean
+	public HiddenHttpMethodFilter hiddenHttpMethodFilter(){
+		return new HiddenHttpMethodFilter();
+	}
 }